+++ Carbony/Modal/Goal.swift
@@ -11,10 +11,10 @@ class Goal: Hashable {
     let uuid: UUID
     let target: Int
     let targetLeft: Int
-    let progress: Double
+    let progress: Int
     let description: String
     
-    init(uuid: UUID, target: Int, targetLeft: Int, progress: Double, description: String) {
+    init(uuid: UUID, target: Int, targetLeft: Int, progress: Int, description: String) {
         self.uuid = uuid
         self.target = target
         self.targetLeft = targetLeft
